refator : move auth beanConfig from domain to infra
