+++ src/main/java/toolc/yourlist/auth/domain/All.java
@@ -1,6 +1,6 @@
 package toolc.yourlist.auth.domain;
 
-class All implements LoginIdPolicy {
+public class All implements LoginIdPolicy {
   private final LoginIdPolicy nonNull = new NonNull();
   private final LoginIdPolicy allowOnlyLowerCaseOrNumber = new AllowOnlyLowerCaseOrNumber();
   private final LoginIdPolicy lengthLimit = new LengthLimit();
