+++ src/main/java/toolc/yourlist/auth/domain/RefreshTokenCreator.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import java.time.Duration;
 
 @RequiredArgsConstructor
-class RefreshTokenCreator {
+public class RefreshTokenCreator {
   private final CurrentTime currentTime;
   private final Duration PC_EXPIRATION_DATE = Duration.ofDays(30);
   private final Duration APP_EXPIRATION_DATE = Duration.ofDays(7);
